requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0

# YOLO
ultralytics>=8.0.0

# LayoutLM
transformers>=4.30.0
datasets>=2.14.0

# OCR
paddleocr>=2.7.0
paddlepaddle>=2.5.0
# easyocr>=1.7.0  # Alternative OCR

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# VLM APIs
openai>=1.0.0
anthropic>=0.18.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging & Monitoring
loguru>=0.7.0
